Wrap tickets, checkout and basket pages in standard html containers     * Add javascript to handle if basket, checkout and complete pages are reached without correct data     * Add basket icon and counter     * Implement function to calculate basket counter and overlay it on the basket icon     * Remove duplicate all-super-min icons file

+++ static/js/complete.js
@@ -4,22 +4,23 @@ const stripe = Stripe(STRIPE_PUBLISHABLE_KEY);
 
 const clientSecret = new URLSearchParams(window.location.search).get("payment_intent_client_secret");
 
-stripe.retrievePaymentIntent(clientSecret).then(({paymentIntent}) => {
-	const message = document.querySelector("#message");
+if(clientSecret)
+	stripe.retrievePaymentIntent(clientSecret).then(({paymentIntent}) => {
+		const message = document.querySelector("#message");
 
-	switch(paymentIntent.status) {
-		case "succeeded":
-			message.innerHTML = "Success! Payment received.";
-			localStorage.clear();
-			break;
-		case "processing":
-			message.innerHTML = "Payment processing. We'll update you when payment is received.";
-			break;
-		case "requires_payment_method":
-			message.innerHTML = "Payment failed. Please try another payment method";
-			break;
-		default:
-			message.innerHTML = "Something went wrong.";
-			break;
-	}
-});
+		switch(paymentIntent.status) {
+			case "succeeded":
+				message.innerHTML = "Success! Payment received.";
+				localStorage.clear();
+				break;
+			case "processing":
+				message.innerHTML = "Payment processing. We'll update you when payment is received.";
+				break;
+			case "requires_payment_method":
+				message.innerHTML = "Payment failed. Please try another payment method";
+				break;
+			default:
+				message.innerHTML = "Something went wrong.";
+				break;
+		}
+	});
